Check faction for GameObject spell casts too.
