+++ game/world/managers/objects/units/UnitManager.py
@@ -999,52 +999,3 @@ class UnitManager(ObjectManager):
     # override
     def generate_object_guid(self, low_guid):
         return low_guid | HighGuid.HIGHGUID_UNIT
-
-    def _allegiance_status_checker(self, target, check_friendly=True):
-        own_faction = DbcDatabaseManager.FactionTemplateHolder.faction_template_get_by_id(self.faction)
-        target_faction = DbcDatabaseManager.FactionTemplateHolder.faction_template_get_by_id(target.faction)
-
-        if not own_faction:
-            Logger.error(f'Invalid faction template: {self.faction}.')
-            return not check_friendly
-
-        if not target_faction:
-            Logger.error(f'Invalid faction template: {target.faction}.')
-            return not check_friendly
-
-        own_enemies = [own_faction.Enemies_1, own_faction.Enemies_2, own_faction.Enemies_3, own_faction.Enemies_4]
-        own_friends = [own_faction.Friend_1, own_faction.Friend_2, own_faction.Friend_3, own_faction.Friend_4]
-        if target_faction.Faction > 0:
-            for enemy in own_enemies:
-                if enemy == target_faction.Faction:
-                    return not check_friendly
-            for friend in own_friends:
-                if friend == target_faction.Faction:
-                    return check_friendly
-
-        if check_friendly:
-            return ((own_faction.FriendGroup & target_faction.FactionGroup) or (own_faction.FactionGroup & target_faction.FriendGroup)) != 0
-        else:
-            return ((own_faction.EnemyGroup & target_faction.FactionGroup) or (own_faction.FactionGroup & target_faction.EnemyGroup)) != 0
-
-    # noinspection PyUnresolvedReferences
-    # override
-    def can_attack_target(self, target):
-        if target is self:
-            return False
-
-        is_enemy = self.is_enemy_to(target)
-        if is_enemy or self.get_type_id() != ObjectTypeIds.ID_PLAYER or target.get_type_id() != ObjectTypeIds.ID_PLAYER:
-            return is_enemy
-
-        if not self.duel_manager:
-            return is_enemy
-
-        # Return true if the players are friendly but dueling
-        return self.duel_manager.player_involved(target) and self.duel_manager.duel_state == DuelState.DUEL_STATE_STARTED
-
-    def is_friendly_to(self, target):
-        return self._allegiance_status_checker(target, True)
-
-    def is_enemy_to(self, target):
-        return self._allegiance_status_checker(target, False)
